Update cc.css
If a gallery caption is of significant length, the flexbox makes the gallery image hilariously small. This is unnecessary. There's probably a better way to do this (ideally, by stripping out all the flex code from zebra-row and related divs, but why do all that when I can just clamp the image size?